Cleaning up Dialects.
